requirements.txt
# Core Scientific Computing
numpy>=1.19.0
pandas>=1.1.0

# Image Processing
opencv-python>=4.5.0
pillow>=8.0.0

# Deep Learning (Uncomment the one you're using)
# TensorFlow
tensorflow>=2.5.0
# OR PyTorch
# torch>=1.9.0
# torchvision>=0.10.0

# Data Science & ML
scikit-learn>=0.24.0
scikit-image>=0.18.0

# Visualization
matplotlib>=3.3.0
seaborn>=0.11.0

# Jupyter & Notebooks
jupyter>=1.0.0
ipython>=7.0.0

# Utilities
tqdm>=4.50.0
python-dotenv>=0.19.0

# Development (optional)
pytest>=6.0.0
black>=21.0.0
flake8>=3.9.0
